Changed application.py -Added user_id variable to dyanamic_jaccard function. Changed geneweaverdb.py -Added function to find and delete jaccards where user_id is true.

+++ src/geneweaverdb.py
@@ -3880,17 +3880,20 @@ def get_genesets_by_hom_id(hom_ids):
     return geneset_list[0]
 
 
-def insert_into_geneset_jaccard(jaccards, gs_id):
+def insert_into_geneset_jaccard(jaccards, gs_id, user_id):
+    # Added user_id to input.
     """
-    First deletes existing geneset jac_values from geneset_jaccard where gs_id is true.
+    First deletes existing geneset jac_values from geneset_jaccard where gs_id is true AND where user_id is true.
+    --This added user_id argument should contain each insert/deletion so that conflicting runs won't fail due to
+    --duplicate add/delete.
     Then insert new values into the table such that gs_id_left < gs_id_right
     :param jaccards: dictionary of jaccard values per gs_id: gs_id => jaccard value
     :param gs_id: gs_id of interest
     :return: 1
     """
-    # delete items from geneset_jaccard
     with PooledCursor() as cursor:
-        cursor.execute('''DELETE FROM geneset_jaccard WHERE gs_id_left=%s OR gs_id_right=%s''', (gs_id, gs_id,))
+        # Overlaid new command onto old script. Now it searches for user_id to delete corresponding rows.
+        cursor.execute('''DELETE FROM geneset_jaccard WHERE gs_id_left=%s OR gs_id_right=%s AND user_id=%s''', (gs_id, gs_id, user_id, ))
         cursor.connection.commit()
         # insert gs_ids left and right
         for key, value in jaccards.items():
@@ -3900,8 +3903,9 @@ def insert_into_geneset_jaccard(jaccards, gs_id):
             else:
                 gs_left = gs_id
                 gs_right = key
-            cursor.execute('''INSERT INTO geneset_jaccard (gs_id_left, gs_id_right, jac_value) VALUES (%s, %s, %s)''',
-                        (gs_left, gs_right, value, ))
+            # Adds back the user_id and does the same as the old script
+            cursor.execute('''INSERT INTO geneset_jaccard (gs_id_left, gs_id_right, jac_value, user_id) VALUES (%s, %s, %s, %s)''',
+                        (gs_left, gs_right, value, user_id, ))
         cursor.connection.commit()
     return 1
